Add tag filters container
