+++ src/styles/custom.css
@@ -44,6 +44,16 @@ button#alert-button:focus {
   background-color: #323232;
 }
 
-button.tag-bar {
+.tag-bar button {
   box-shadow: 0 0 2px 0 black;
+}
+
+.tag-bar .list-group {
+  overflow-x: auto;
+  overflow-y: hidden;
+  height: 3rem;
+  max-height: 3rem;
+}
+.tag-bar .list-group-item {
+  white-space: nowrap;
 }
